Donation Pledge MVP
+ Migration to add pledge related fields for fundraisers table
+ Don't settle payments for a pledger
+ Add form fields to UI
