+++ app/models/page.rb
@@ -127,6 +127,16 @@ class Page < ActiveRecord::Base
                                end
   end
 
+  def pledger?
+    @fundraiser ||= Plugins::Fundraiser.find_by(page_id: id)
+
+    @fundraiser.try(:pledge) || false
+  end
+
+  def set_as_pledger
+    Plugins::Fundraiser.find_by(page_id: id).set_as_pledger
+  end
+
   private
 
   def switch_plugins
