+++ db/migrate/20161207140508_add_pledge_flag_top_plugins_fundraisers.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+class AddPledgeFlagTopPluginsFundraisers < ActiveRecord::Migration
+  def change
+    add_column :plugins_fundraisers, :pledge, :boolean, default: false
+    add_column :plugins_fundraisers, :pledge_processed_on, :datetime
+    add_column :plugins_fundraisers, :pledge_target_in_actions, :integer, default: 0
+  end
+end
