+++ db/migrate/20161212124035_add_plege_flag_to_payment_braintree_transaction.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+class AddPlegeFlagToPaymentBraintreeTransaction < ActiveRecord::Migration
+  def change
+    add_column :payment_braintree_transactions, :pledge, :boolean, default: false
+    add_column :payment_braintree_transactions, :pledge_processed_at, :datetime
+  end
+end
