+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20161103092722) do
+ActiveRecord::Schema.define(version: 20161212124035) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -305,8 +305,8 @@ ActiveRecord::Schema.define(version: 20161103092722) do
     t.datetime "transaction_created_at"
     t.string   "payment_method_token"
     t.string   "customer_id"
-    t.datetime "created_at",                                       null: false
-    t.datetime "updated_at",                                       null: false
+    t.datetime "created_at",                                                       null: false
+    t.datetime "updated_at",                                                       null: false
     t.string   "merchant_account_id"
     t.string   "currency"
     t.integer  "page_id"
@@ -316,6 +316,8 @@ ActiveRecord::Schema.define(version: 20161103092722) do
     t.string   "processor_response_code"
     t.integer  "payment_method_id"
     t.integer  "subscription_id"
+    t.boolean  "pledge",                                           default: false
+    t.datetime "pledge_processed_at"
   end
 
   add_index "payment_braintree_transactions", ["page_id"], name: "index_payment_braintree_transactions_on_page_id", using: :btree
@@ -412,12 +414,15 @@ ActiveRecord::Schema.define(version: 20161103092722) do
     t.string   "title"
     t.string   "ref"
     t.integer  "page_id"
-    t.boolean  "active",            default: false
-    t.datetime "created_at",                        null: false
-    t.datetime "updated_at",                        null: false
+    t.boolean  "active",                   default: false
+    t.datetime "created_at",                               null: false
+    t.datetime "updated_at",                               null: false
     t.integer  "form_id"
     t.integer  "donation_band_id"
-    t.integer  "recurring_default", default: 0,     null: false
+    t.integer  "recurring_default",        default: 0,     null: false
+    t.boolean  "pledge",                   default: false
+    t.datetime "pledge_processed_on"
+    t.integer  "pledge_target_in_actions", default: 0
   end
 
   add_index "plugins_fundraisers", ["donation_band_id"], name: "index_plugins_fundraisers_on_donation_band_id", using: :btree
